devide function
